optimize stat calls in resolvers
