+++ src/index.ts
@@ -17,6 +17,7 @@ import {
   parse,
   ProjectLocalResolveHelper,
   split,
+  versionGteLt,
   yn,
 } from './util';
 import { findAndReadConfig, loadCompiler } from './configuration';
@@ -65,33 +66,6 @@ export type {
  */
 const engineSupportsPackageTypeField =
   parseInt(process.versions.node.split('.')[0], 10) >= 12;
-
-/** @internal */
-export function versionGteLt(
-  version: string,
-  gteRequirement: string,
-  ltRequirement?: string
-) {
-  const [major, minor, patch, extra] = parse(version);
-  const [gteMajor, gteMinor, gtePatch] = parse(gteRequirement);
-  const isGte =
-    major > gteMajor ||
-    (major === gteMajor &&
-      (minor > gteMinor || (minor === gteMinor && patch >= gtePatch)));
-  let isLt = true;
-  if (ltRequirement) {
-    const [ltMajor, ltMinor, ltPatch] = parse(ltRequirement);
-    isLt =
-      major < ltMajor ||
-      (major === ltMajor &&
-        (minor < ltMinor || (minor === ltMinor && patch < ltPatch)));
-  }
-  return isGte && isLt;
-
-  function parse(requirement: string) {
-    return requirement.split(/[\.-]/).map((s) => parseInt(s, 10));
-  }
-}
 
 /**
  * Assert that script can be loaded as CommonJS when we attempt to require it.
